Clarify what is documenting the process
The library is not documenting the process; the guide is.  FormatJS is not providing the JavaScript library; FormatJS is the library.  

I chose to focus on just the guide in this section.  Alternatively, we could write: 

Internationalizing web apps is an involved and complex task.  We created {{brand}} to make internationalization simple.  We created this guide to inform you about the whole internationalization context.