Revamp build to de-dup intl integrations code
This website is odd in that it requires all of FormatJS' integration
libraries, which meant tons of duplicate code shared between all of the
integration libs was loaded on each page.

This revamps the build process to de-dup the shared code between
FormatJS' high-level integration libs together by building the ES6
modules. It also creates a builds a `locale-data.js` file that
contains the data for the locales this site supports.

Overall this change removes one HTTP request and ~100KB of JavaScript.

Relates to #71