Allow examples to work with a sub-set of locales
This updates the site's examples infrastructure so that some examples
can work even if they don't support every locale the whole site works
with.

This makes it possible to add new message formatting examples in
English before waiting for someone to property translate the example to
every locale we want the site to support.