Include locale data for each integration
This updates the Broccoli build to include the locale data for each
integration. Doing this makes sure that if the lib's deps get out of
sync that they all still propagate the locale data down to the copy of
`IntlMessageFormat` and `IntlRelativeFormat` they depend on.

This adds bulk to the `locale-data.js` file that's created, but it's
**highly** compressible with gzip since the data is all duplicated.