Use an updating relative time in home page example
This switches the home page "splash" example to use a relative time,
and also forces it to re-render every second so the time keeps on
climbing while someone is on the page. It makes the example more
complex and shows off the power of FormatJS better.

_Note: The other locales will need a `photosNested` message that
matches the format of the `en-US` default one._