Hack Istanbul to ignore es6-shim
Istanbul was instrumenting the `es6-shim` module because it's a
polyfill that provides global functions which are invoked by the unit
tests.

I was unable to configure Istanbul to ignore this file, so I had to
misuse the `--post-require-hook` CLI option to load this "patch". The
hack takes control of `istanbul.hook.hookRequire` to opt-out of
transformation if a file matches one of the ignored files.