Replace Example mixin with composable components
This replaces the `Example` mixin used by the various `<*Example>`
components with a new high-level component: `<ExampleContainer>`, and
a new low-level `<Example>` component.

These two new components are used with the `<HandlebarsExample>`,
`<ReactExample>`, etc. components to keep thing DRY, and also allow
for everything _but_ the `<ExampleContainer>` to be pure components (
i.e., they don't have `state`.)

Now `<ExampleContainer>` contains the state and logic for dealing with
an example's `currentLocale`, and once it has the data ready it
delegates rendering to the specialized `<*Example>` components for each
type of integration.

The new component hierarchy looks like this:

    <ExampleContainer {...props} component={HandlebarsExample} />

    // The `render()` in <HandlebarsExample> looks like this:

    <Example {...props}
         source={<Tabs>{tabs}</Tabs>}
         output={<HandlebarsOutput {...props} />} />